refactor: map
